change files & add more metadata
