+++ app/account/frontpage/page.jsx
@@ -2,6 +2,10 @@
 import Layout from "@/app/layouts/account-layout";
 import FrontPage from "@/components/account/frontPage";
 
+export const metadata = {
+  title: "Account",
+};
+
 export default function AccountPage() {
   return (
     <>
