+++ app/account/gallery/[uniqueId]/page.jsx
@@ -4,6 +4,10 @@ import Layout from "@/app/layouts/account-layout";
 
 export const runtime = "edge";
 
+export const metadata = {
+  title: "Account Gallery",
+};
+
 const ImageAccountTools = () => {
   return (
     <Layout>
