+++ app/account/gallery/page.jsx
@@ -4,6 +4,10 @@ import AccountGallery from "@/components/account/gallery/accountGallery";
 
 export const runtime = "edge";
 
+export const metadata = {
+  title: "Account Gallery",
+};
+
 export default function FetchGallery() {
 
   return (
