+++ app/account/page.jsx
@@ -2,6 +2,10 @@
 import Layout from "@/app/layouts/account-layout";
 import AccountPage from "@/components/account/accountPage";
 
+export const metadata = {
+  title: "Account Settings",
+};
+
 export default function AccountSettings() {
   return (
     <>
