+++ app/account/upload/page.jsx
@@ -2,6 +2,10 @@
 import Layout from "@/app/layouts/account-layout";
 import Upload from "@/components/account/upload";
 
+export const metadata = {
+  title: "Upload",
+};
+
 export default function UploadPage() {
   return (
     <>
