+++ app/gallery/[uniqueId]/page.jsx
