+++ app/legal/page.jsx
@@ -1,8 +1,11 @@
-"use client";
 import { PaperClipIcon } from "@heroicons/react/20/solid";
 import Header from "@/components/header";
 import Footer from "@/components/footer";
 import Link from "next/link";
+
+export const metadata = {
+  title: "Legal",
+};
 
 export default function LegalPage() {
   return (
