+++ app/login/page.jsx
@@ -1,8 +1,11 @@
-'use client';
 import Header from "@/components/header";
 import Login from "@/components/login";
 
-export default function Gallery() {
+export const metadata = {
+  title: "Login",
+};
+
+export default function LoginPage() {
 
     return (
       <div>
