+++ components/public/fetchGallerySingle.jsx
