+++ components/public/fetchPublicGallery.jsx
