+++ components/user/galleryPage.jsx
